app/lib/duty_schedule.py
import datetime
import pickle
import os
from googleapiclient.discovery import build
from google_auth_oauthlib.flow import InstalledAppFlow
from google.auth.transport.requests import Request
import logging


log = logging.getLogger('default')

# If modifying these scopes, delete the file token.pickle.
SCOPES = ['https://www.googleapis.com/auth/calendar.readonly']


def get_agent():

    token_file = '/etc/app/token.pickle'
    #credential_file = '/etc/app/credentials.json'
    dba_duty_cal = os.getenv('dba_duty_cal')

    #today = datetime.datetime.now().date().isoformat()
    creds = None
    # The fil.pickle stores the user's access and refresh tokens, and is
    # created automatically when the authorization flow completes for the first
    # time.
    if os.path.exists(token_file):
        with open(token_file, 'rb') as token:
            creds = pickle.load(token)
    # If there are no (valid) credentials available, let the user log in.
    if not creds or not creds.valid:
        if creds and creds.expired and creds.refresh_token:
            creds.refresh(Request())
        else:
            flow = InstalledAppFlow.from_client_secrets_file(
                token_file, SCOPES)
            creds = flow.run_local_server(port=0)
        # Save the credentials for the next run
        with open(token_file, 'wb') as token:
            pickle.dump(creds, token)

    service = build('calendar', 'v3', credentials=creds)

    # Call the Calendar API
    now = datetime.datetime.utcnow().isoformat() + 'Z' # 'Z' indicates UTC time
    events_result = service.events().list(calendarId=dba_duty_cal, timeMin=now,
                                        maxResults=1, singleEvents=True,
                                        orderBy='startTime').execute()
    events = events_result.get('items', [])

    agent = None
    if not events:
        log.warning('No on-call agent found.')
    else:
 #       for event in events:
 #           pass
        event = events[0]  # only support one event(agent name)
        agent = event['summary']
        log.info('Found on-call agent %s.' % agent)
    return agent
